move chi middleware
